+++ packages/RSuite/NEWS.md
@@ -1,3 +1,8 @@
+# RSuite 0.30 (2018-07-??)
+  * RSuite package supports MacOS from now on: proper repository url handling,
+     commands to download/build packages adapted to run on MacOS, handling of
+     mac.binary type added.
+
 # RSuite 0.29 (2018-06-24)
   * #169: if project package depends on some dependents of devtools and devtools
      is installed into sbox as supportive package not available in the